TIGER batch correction

This function returns a pmart object that has been undergone TIGER batch effect correction

bc_tiger(
  omicsData,
  sampletype_cname,
  test_val,
  position_cname = NULL,
  injection_cname = NULL
)

Arguments

omicsData

an object of the class 'pepData', 'proData', 'metabData', 'lipidData', or 'nmrData', usually created by as.pepData, as.proData, as.metabData, as.lipidData, or as.nmrData, respectively.

sampletype_cname

character string giving the name of the column in omicsData$f_data that contains the sample type information (such as quality control samples)

test_val

character string giving the name of the value within the column sampletype_cname to be used as the testing value for TIGER

position_cname

character string giving the name of the column in omicsData$f_data that contains the well position, default is NULL

injection_cname

character string giving the name of the column in omicsData$f_data that contains the injection order, default is NULL

Value

Object of same class as omicsData that has been undergone TIGER normalization

Author

Damon Leach

Examples

library(malbacR)
library(pmartR)
data("pmart_amide")
pmart_amide <- edata_transform(pmart_amide,"log2")
pmart_amide <- group_designation(pmart_amide,main_effects = "group",batch_id = "batch")
pmart_amide <- normalize_global(pmart_amide,subset_fn = "all",norm_fn = "median",
                               apply_norm = TRUE,backtransform = TRUE)
tigerFilt <- tiger_filter(pmart_amide,sampletype_cname = "group",test_val = "QC")
pmart_amideFilt <- apply_tigerFilt(tigerFilt,pmart_amide)
#> There are no molecules that need to be removed. Returning the original omicsData
amide_tiger <- bc_tiger(omicsData = pmart_amideFilt,sampletype_cname = "group",
                        test_val = "QC",injection_cname = "Injection_order")
#> + Initialising...   2023-10-17 12:24:11
#> + Computing target values...   2023-10-17 12:24:11
#> + Selecting highly-correlated variables...   2023-10-17 12:24:11
#>   - Computing correlation coefficients...
#>   - Selecting variables...
#> + Data correction started.   2023-10-17 12:24:12
#>   |                                                                      | 0 % ~calculating    |==                                                                    | 2 % ~03m 14s        |===                                                                   | 4 % ~03m 07s        |====                                                                  | 5 % ~03m 01s        |=====                                                                 | 7 % ~02m 55s        |=======                                                               | 9 % ~02m 52s        |========                                                              | 11% ~02m 47s        |=========                                                             | 12% ~02m 44s        |==========                                                            | 14% ~02m 41s        |============                                                          | 16% ~02m 37s        |=============                                                         | 18% ~02m 33s        |==============                                                        | 20% ~02m 30s        |===============                                                       | 21% ~02m 26s        |=================                                                     | 23% ~02m 24s        |==================                                                    | 25% ~02m 20s        |===================                                                   | 27% ~02m 17s        |====================                                                  | 29% ~02m 14s        |======================                                                | 30% ~02m 11s        |=======================                                               | 32% ~02m 07s        |========================                                              | 34% ~02m 04s        |=========================                                             | 36% ~02m 00s        |===========================                                           | 38% ~01m 57s        |============================                                          | 39% ~01m 53s        |=============================                                         | 41% ~01m 50s        |==============================                                        | 43% ~01m 46s        |================================                                      | 45% ~01m 43s        |=================================                                     | 46% ~01m 39s        |==================================                                    | 48% ~01m 36s        |===================================                                   | 50% ~01m 33s        |=====================================                                 | 52% ~01m 30s        |======================================                                | 54% ~01m 26s        |=======================================                               | 55% ~01m 23s        |========================================                              | 57% ~01m 20s        |==========================================                            | 59% ~01m 16s        |===========================================                           | 61% ~01m 13s        |============================================                          | 62% ~01m 10s        |=============================================                         | 64% ~01m 06s        |===============================================                       | 66% ~01m 03s        |================================================                      | 68% ~60s            |=================================================                     | 70% ~57s            |==================================================                    | 71% ~53s            |====================================================                  | 73% ~50s            |=====================================================                 | 75% ~47s            |======================================================                | 77% ~43s            |=======================================================               | 79% ~40s            |=========================================================             | 80% ~37s            |==========================================================            | 82% ~33s            |===========================================================           | 84% ~30s            |============================================================          | 86% ~27s            |==============================================================        | 88% ~23s            |===============================================================       | 89% ~20s            |================================================================      | 91% ~17s            |=================================================================     | 93% ~13s            |===================================================================   | 95% ~10s            |====================================================================  | 96% ~07s            |===================================================================== | 98% ~03s            |======================================================================| 100% elapsed=03m 08s
#>   - Merging results...
#> + Completed.   2023-10-17 12:27:21