WaveICA_2.0 batch correction

This function returns a pmart object that has been undergone WaveICA_2.0 batch effect correction

bc_waveica(omicsData, injection_cname, alpha = 0, cutoff = 0.1, K = 10)

Arguments

omicsData

an object of the class 'pepData', 'proData', 'metabData', 'lipidData', or 'nmrData', usually created by as.pepData, as.proData, as.metabData, as.lipidData, or as.nmrData, respectively.

injection_cname

character string giving the name of the column in omicsData$f_data that contains the run order of the samples

alpha

tradeoff value between the independence of samples and those of variables in ICA, and should be between 0 and 1, defaults to 0

cutoff

threshold of the variation explained by the injection order for independent components, should be between 0 and 1, defaults to 0.1

K

maximal component that ICA decomposes, defaults to 10

Value

Object of same class as omicsData that has been undergone WaveICA normalization

Author

Damon Leach

Examples

library(malbacR)
library(pmartR)
data("pmart_amide")
pmart_amide <- edata_transform(pmart_amide,"log2")
pmart_amide <- group_designation(pmart_amide,main_effects = "group",batch_id = "batch")
pmart_amide <- normalize_global(pmart_amide,subset_fn = "all",norm_fn = "median",
                               apply_norm = TRUE,backtransform = TRUE)
impObj <- imputation(omicsData = pmart_amide)
amide_imp <- apply_imputation(imputeData = impObj, omicsData = pmart_amide)
amide_wave <- bc_waveica(omicsData = amide_imp, injection_cname = "Injection_order",
                         alpha = 0, cutoff = 0.1, K = 10)
#> 
#> waveslim: Wavelet Method for 1/2/3D Signals (version = 1.8.4)
#> Loading required package: nlme
#> This is mgcv 1.8-41. For overview type 'help("mgcv-package")'.
#> 
#> Attaching package: ‘mgcv’
#> The following object is masked from ‘package:ica’:
#> 
#>     sdiag
#> ######Decomposition 1 ########
#> ######Decomposition 2 ########
#> ######Decomposition 3 ########
#> ######Decomposition 4 ########
#> ######Decomposition 5 ########
#> ######Decomposition 6 ########
#> ######Decomposition 7 ########
#> ######Decomposition 8 ########
#> ######Decomposition 9 ########
#> ######Decomposition 10 ########
#> ######Decomposition 11 ########
#> ######Decomposition 12 ########
#> ######Decomposition 13 ########
#> ######Decomposition 14 ########
#> ######Decomposition 15 ########
#> ######Decomposition 16 ########
#> ######Decomposition 17 ########
#> ######Decomposition 18 ########
#> ######Decomposition 19 ########
#> ######Decomposition 20 ########
#> ######Decomposition 21 ########
#> ######Decomposition 22 ########
#> ######Decomposition 23 ########
#> ######Decomposition 24 ########
#> ######Decomposition 25 ########
#> ######Decomposition 26 ########
#> ######Decomposition 27 ########
#> ######Decomposition 28 ########
#> ######Decomposition 29 ########
#> ######Decomposition 30 ########
#> ######Decomposition 31 ########
#> ######Decomposition 32 ########
#> ######Decomposition 33 ########
#> ######Decomposition 34 ########
#> ######Decomposition 35 ########
#> ######Decomposition 36 ########
#> ######Decomposition 37 ########
#> ######Decomposition 38 ########
#> ######Decomposition 39 ########
#> ######Decomposition 40 ########
#> ######Decomposition 41 ########
#> ######Decomposition 42 ########
#> ######Decomposition 43 ########
#> ######Decomposition 44 ########
#> ######Decomposition 45 ########
#> ######Decomposition 46 ########
#> ######Decomposition 47 ########
#> ######Decomposition 48 ########
#> ######Decomposition 49 ########
#> ######Decomposition 50 ########
#> ######Decomposition 51 ########
#> ######Decomposition 52 ########
#> ######Decomposition 53 ########
#> ######Decomposition 54 ########
#> ######Decomposition 55 ########
#> ######Decomposition 56 ########
#> ######Decomposition 57 ########
#> ######Decomposition 58 ########
#> ######Decomposition 59 ########
#> ######Decomposition 60 ########
#> ######Decomposition 61 ########
#> ######Decomposition 62 ########
#> ######Decomposition 63 ########
#> ######Decomposition 64 ########
#> ######Decomposition 65 ########
#> ######Decomposition 66 ########
#> ######Decomposition 67 ########
#> ######Decomposition 68 ########
#> ######Decomposition 69 ########
#> ######Decomposition 70 ########
#> ######Decomposition 71 ########
#> ######Decomposition 72 ########
#> ######Decomposition 73 ########
#> ######Decomposition 74 ########
#> ######Decomposition 75 ########
#> ######Decomposition 76 ########
#> ######Decomposition 77 ########
#> ######Decomposition 78 ########
#> ######Decomposition 79 ########
#> ######Decomposition 80 ########
#> ######Decomposition 81 ########
#> ######Decomposition 82 ########
#> ######Decomposition 83 ########
#> ######Decomposition 84 ########
#> ######Decomposition 85 ########
#> ######Decomposition 86 ########
#> ######Decomposition 87 ########
#> ######Decomposition 88 ########
#> ######Decomposition 89 ########
#> ######Decomposition 90 ########
#> ######Decomposition 91 ########
#> ######Decomposition 92 ########
#> ######Decomposition 93 ########
#> ######Decomposition 94 ########
#> ######Decomposition 95 ########
#> ######Decomposition 96 ########
#> ######Decomposition 97 ########
#> ######Decomposition 98 ########
#> ######Decomposition 99 ########
#> ######Decomposition 100 ########
#> ######Decomposition 101 ########
#> ######Decomposition 102 ########
#> ######Decomposition 103 ########
#> ######Decomposition 104 ########
#> ######Decomposition 105 ########
#> ######Decomposition 106 ########
#> ######Decomposition 107 ########
#> ######Decomposition 108 ########
#> ######Decomposition 109 ########
#> ######Decomposition 110 ########
#> ######Decomposition 111 ########
#> ######Decomposition 112 ########
#> ######Decomposition 113 ########
#> ######Decomposition 114 ########
#> ######Decomposition 115 ########
#> ######Decomposition 116 ########
#> ######Decomposition 117 ########
#> ######Decomposition 118 ########
#> ######Decomposition 119 ########
#> ######Decomposition 120 ########
#> ######Decomposition 121 ########
#> ######Decomposition 122 ########
#> ######Decomposition 123 ########
#> ######Decomposition 124 ########
#> ######Decomposition 125 ########
#> ######Decomposition 126 ########
#> ######Decomposition 127 ########
#> ######Decomposition 128 ########
#> ######Decomposition 129 ########
#> ######Decomposition 130 ########
#> ######Decomposition 131 ########
#> ######Decomposition 132 ########
#> ######Decomposition 133 ########
#> ######Decomposition 134 ########
#> ######Decomposition 135 ########
#> ######Decomposition 136 ########
#> ######Decomposition 137 ########
#> ######Decomposition 138 ########
#> ######Decomposition 139 ########
#> ######Decomposition 140 ########
#> ######Decomposition 141 ########
#> ######Decomposition 142 ########
#> ######Decomposition 143 ########
#> ######Decomposition 144 ########
#> ######Decomposition 145 ########
#> ######Decomposition 146 ########
#> ######Decomposition 147 ########
#> ######Decomposition 148 ########
#> ######Decomposition 149 ########
#> ######Decomposition 150 ########
#> ######Decomposition 151 ########
#> ######Decomposition 152 ########
#> ######Decomposition 153 ########
#> ######Decomposition 154 ########
#> ######Decomposition 155 ########
#> ######Decomposition 156 ########
#> ######Decomposition 157 ########
#> ######Decomposition 158 ########
#> ######Decomposition 159 ########
#> ######Decomposition 160 ########
#> ######Decomposition 161 ########
#> ######Decomposition 162 ########
#> ######Decomposition 163 ########
#> ######Decomposition 164 ########
#> ######Decomposition 165 ########
#> ######Decomposition 166 ########
#> ######Decomposition 167 ########
#> ######Decomposition 168 ########
#> ######Decomposition 169 ########
#> ######Decomposition 170 ########
#> ######Decomposition 171 ########
#> ######Decomposition 172 ########
#> ######Decomposition 173 ########
#> ######Decomposition 174 ########
#> ######Decomposition 175 ########
#> ######Decomposition 176 ########
#> ######Decomposition 177 ########
#> ######Decomposition 178 ########
#> ######Decomposition 179 ########
#> ######Decomposition 180 ########
#> ######Decomposition 181 ########
#> ######Decomposition 182 ########
#> ######Decomposition 183 ########
#> ######Decomposition 184 ########
#> ######Decomposition 185 ########
#> ######Decomposition 186 ########
#> ######Decomposition 187 ########
#> ######Decomposition 188 ########
#> ######Decomposition 189 ########
#> ######Decomposition 190 ########
#> ######Decomposition 191 ########
#> ######Decomposition 192 ########
#> ######Decomposition 193 ########
#> ######Decomposition 194 ########
#> ######Decomposition 195 ########
#> ######Decomposition 196 ########
#> ######Decomposition 197 ########
#> ######Decomposition 198 ########
#> ######Decomposition 199 ########
#> ######Decomposition 200 ########
#> ######Decomposition 201 ########
#> ######Decomposition 202 ########
#> ######Decomposition 203 ########
#> ######Decomposition 204 ########
#> ######Decomposition 205 ########
#> ######Decomposition 206 ########
#> ######Decomposition 207 ########
#> ######Decomposition 208 ########
#> ######Decomposition 209 ########
#> ######Decomposition 210 ########
#> ######Decomposition 211 ########
#> ######Decomposition 212 ########
#> ######Decomposition 213 ########
#> ######Decomposition 214 ########
#> ######Decomposition 215 ########
#> ######Decomposition 216 ########
#> ######Decomposition 217 ########
#> ######Decomposition 218 ########
#> ######Decomposition 219 ########
#> ######Decomposition 220 ########
#> ######Decomposition 221 ########
#> ######Decomposition 222 ########
#> ######Decomposition 223 ########
#> ######Decomposition 224 ########
#> ######Decomposition 225 ########
#> ######Decomposition 226 ########
#> ######Decomposition 227 ########
#> ######Decomposition 228 ########
#> ######Decomposition 229 ########
#> ######Decomposition 230 ########
#> ######Decomposition 231 ########
#> ######Decomposition 232 ########
#> ######Decomposition 233 ########
#> ######Decomposition 234 ########
#> ######Decomposition 235 ########
#> ######Decomposition 236 ########
#> ######Decomposition 237 ########
#> ######Decomposition 238 ########
#> ######Decomposition 239 ########
#> ######Decomposition 240 ########
#> ######Decomposition 241 ########
#> ######Decomposition 242 ########
#> ######Decomposition 243 ########
#> ######Decomposition 244 ########
#> ######Decomposition 245 ########
#> ######Decomposition 246 ########
#> ######Decomposition 247 ########
#> ######Decomposition 248 ########
#> ######Decomposition 249 ########
#> ######Decomposition 250 ########
#> ######Decomposition 251 ########
#> ######Decomposition 252 ########
#> ######Decomposition 253 ########
#> ######Decomposition 254 ########
#> ######Decomposition 255 ########
#> ######Decomposition 256 ########
#> ######Decomposition 257 ########
#> ######Decomposition 258 ########
#> ######Decomposition 259 ########
#> ######Decomposition 260 ########
#> ######Decomposition 261 ########
#> ######Decomposition 262 ########
#> ######Decomposition 263 ########
#> ######Decomposition 264 ########
#> ######Decomposition 265 ########
#> ######Decomposition 266 ########
#> ######Decomposition 267 ########
#> ######Decomposition 268 ########
#> ######Decomposition 269 ########
#> ######Decomposition 270 ########
#> ######Decomposition 271 ########
#> ######Decomposition 272 ########
#> ######Decomposition 273 ########
#> ######Decomposition 274 ########
#> ######Decomposition 275 ########
#> ######Decomposition 276 ########
#> ######Decomposition 277 ########
#> ######Decomposition 278 ########
#> ######Decomposition 279 ########
#> ######Decomposition 280 ########
#> ######Decomposition 281 ########
#> ######Decomposition 282 ########
#> ######Decomposition 283 ########
#> ######Decomposition 284 ########
#> ######Decomposition 285 ########
#> ######Decomposition 286 ########
#> ######Decomposition 287 ########
#> ######Decomposition 288 ########
#> ######Decomposition 289 ########
#> ######Decomposition 290 ########
#> ######Decomposition 291 ########
#> ######Decomposition 292 ########
#> ######Decomposition 293 ########
#> ######Decomposition 294 ########
#> ######Decomposition 295 ########
#> ######Decomposition 296 ########
#> ######Decomposition 297 ########
#> ######Decomposition 298 ########
#> ######Decomposition 299 ########
#> ######Decomposition 300 ########
#> ######Decomposition 301 ########
#> ######Decomposition 302 ########
#> ######Decomposition 303 ########
#> ######Decomposition 304 ########
#> ######Decomposition 305 ########
#> ######Decomposition 306 ########
#> ######Decomposition 307 ########
#> ######Decomposition 308 ########
#> ######Decomposition 309 ########
#> ######Decomposition 310 ########
#> ######Decomposition 311 ########
#> ######Decomposition 312 ########
#> ######Decomposition 313 ########
#> ######Decomposition 314 ########
#> ######Decomposition 315 ########
#> ######Decomposition 316 ########
#> ######Decomposition 317 ########
#> ######Decomposition 318 ########
#> ######Decomposition 319 ########
#> ######Decomposition 320 ########
#> ######Decomposition 321 ########
#> ######Decomposition 322 ########
#> ######Decomposition 323 ########
#> ######Decomposition 324 ########
#> ######Decomposition 325 ########
#> ######Decomposition 326 ########
#> ######Decomposition 327 ########
#> ######Decomposition 328 ########
#> ######Decomposition 329 ########
#> ######Decomposition 330 ########
#> ######Decomposition 331 ########
#> ######Decomposition 332 ########
#> ######Decomposition 333 ########
#> ######Decomposition 334 ########
#> ######Decomposition 335 ########
#> ######Decomposition 336 ########
#> ######Decomposition 337 ########
#> ######Decomposition 338 ########
#> ######Decomposition 339 ########
#> ######Decomposition 340 ########
#> ######Decomposition 341 ########
#> ######Decomposition 342 ########
#> ######Decomposition 343 ########
#> ######Decomposition 344 ########
#> ######Decomposition 345 ########
#> ######Decomposition 346 ########
#> ######Decomposition 347 ########
#> ######Decomposition 348 ########
#> ######Decomposition 349 ########
#> ######Decomposition 350 ########
#> ######Decomposition 351 ########
#> ######Decomposition 352 ########
#> ######Decomposition 353 ########
#> ######Decomposition 354 ########
#> ######Decomposition 355 ########
#> ######Decomposition 356 ########
#> ######Decomposition 357 ########
#> ######Decomposition 358 ########
#> ######Decomposition 359 ########
#> ######Decomposition 360 ########
#> ######Decomposition 361 ########
#> ######Decomposition 362 ########
#> ######Decomposition 363 ########
#> ######Decomposition 364 ########
#> ######Decomposition 365 ########
#> ######Decomposition 366 ########
#> ######Decomposition 367 ########
#> ######Decomposition 368 ########
#> ######Decomposition 369 ########
#> ######Decomposition 370 ########
#> ######Decomposition 371 ########
#> ######Decomposition 372 ########
#> ######Decomposition 373 ########
#> ######Decomposition 374 ########
#> ######Decomposition 375 ########
#> ######Decomposition 376 ########
#> ######Decomposition 377 ########
#> ######Decomposition 378 ########
#> ######Decomposition 379 ########
#> ######Decomposition 380 ########
#> ######Decomposition 381 ########
#> ######Decomposition 382 ########
#> ######Decomposition 383 ########
#> ######Decomposition 384 ########
#> ######Decomposition 385 ########
#> ######Decomposition 386 ########
#> ######Decomposition 387 ########
#> ######Decomposition 388 ########
#> ######Decomposition 389 ########
#> ######Decomposition 390 ########
#> ######Decomposition 391 ########
#> ######Decomposition 392 ########
#> ######Decomposition 393 ########
#> ######Decomposition 394 ########
#> ######Decomposition 395 ########
#> ######Decomposition 396 ########
#> ######Decomposition 397 ########
#> ######Decomposition 398 ########
#> ######Decomposition 399 ########
#> ######Decomposition 400 ########
#> ######Decomposition 401 ########
#> ######Decomposition 402 ########
#> ######Decomposition 403 ########
#> ######Decomposition 404 ########
#> ######Decomposition 405 ########
#> ######Decomposition 406 ########
#> ######Decomposition 407 ########
#> ######Decomposition 408 ########
#> ######Decomposition 409 ########
#> ######Decomposition 410 ########
#> ######Decomposition 411 ########
#> ######Decomposition 412 ########
#> ######Decomposition 413 ########
#> ######Decomposition 414 ########
#> ######Decomposition 415 ########
#> ######Decomposition 416 ########
#> ######Decomposition 417 ########
#> ######Decomposition 418 ########
#> ######Decomposition 419 ########
#> ######Decomposition 420 ########
#> ######Decomposition 421 ########
#> ######Decomposition 422 ########
#> ######Decomposition 423 ########
#> ######Decomposition 424 ########
#> ######Decomposition 425 ########
#> ######Decomposition 426 ########
#> ######Decomposition 427 ########
#> ######Decomposition 428 ########
#> ######Decomposition 429 ########
#> ######Decomposition 430 ########
#> ######Decomposition 431 ########
#> ######Decomposition 432 ########
#> ######Decomposition 433 ########
#> ######Decomposition 434 ########
#> ######Decomposition 435 ########
#> ######Decomposition 436 ########
#> ######Decomposition 437 ########
#> ######Decomposition 438 ########
#> ######Decomposition 439 ########
#> ######Decomposition 440 ########
#> ######Decomposition 441 ########
#> ######Decomposition 442 ########
#> ######Decomposition 443 ########
#> ######Decomposition 444 ########
#> ######Decomposition 445 ########
#> ######Decomposition 446 ########
#> ######Decomposition 447 ########
#> ######Decomposition 448 ########
#> ######Decomposition 449 ########
#> ######Decomposition 450 ########
#> ######Decomposition 451 ########
#> ######Decomposition 452 ########
#> ######Decomposition 453 ########
#> ######Decomposition 454 ########
#> ######Decomposition 455 ########
#> ######Decomposition 456 ########
#> ######Decomposition 457 ########
#> ######Decomposition 458 ########
#> ######Decomposition 459 ########
#> ######Decomposition 460 ########
#> ######Decomposition 461 ########
#> ######Decomposition 462 ########
#> ######Decomposition 463 ########
#> ######Decomposition 464 ########
#> ######Decomposition 465 ########
#> ######Decomposition 466 ########
#> ######Decomposition 467 ########
#> ######Decomposition 468 ########
#> ######Decomposition 469 ########
#> ######Decomposition 470 ########
#> ######Decomposition 471 ########
#> ######Decomposition 472 ########
#> ######Decomposition 473 ########
#> ######Decomposition 474 ########
#> ######Decomposition 475 ########
#> ######Decomposition 476 ########
#> ######Decomposition 477 ########
#> ######Decomposition 478 ########
#> ######Decomposition 479 ########
#> ######Decomposition 480 ########
#> ######Decomposition 481 ########
#> ######Decomposition 482 ########
#> ######Decomposition 483 ########
#> ######Decomposition 484 ########
#> ######Decomposition 485 ########
#> ######Decomposition 486 ########
#> ######Decomposition 487 ########
#> ######Decomposition 488 ########
#> ######Decomposition 489 ########
#> ######Decomposition 490 ########
#> ######Decomposition 491 ########
#> ######Decomposition 492 ########
#> ######Decomposition 493 ########
#> ######Decomposition 494 ########
#> ######Decomposition 495 ########
#> ######Decomposition 496 ########
#> ######Decomposition 497 ########
#> ######Decomposition 498 ########
#> ######Decomposition 499 ########
#> ######Decomposition 500 ########
#> ######### ICA 1 #############
#> ######### ICA 2 #############
#> ######### ICA 3 #############
#> ######### ICA 4 #############
#> ######### ICA 5 #############
#> ######### ICA 6 #############
#> ######### ICA 7 #############
#> ######### ICA 8 #############
#> ######### ICA 9 #############
#> ######### ICA 10 #############
#> ######Reconstruction 1 ########
#> ######Reconstruction 2 ########
#> ######Reconstruction 3 ########
#> ######Reconstruction 4 ########
#> ######Reconstruction 5 ########
#> ######Reconstruction 6 ########
#> ######Reconstruction 7 ########
#> ######Reconstruction 8 ########
#> ######Reconstruction 9 ########
#> ######Reconstruction 10 ########
#> ######Reconstruction 11 ########
#> ######Reconstruction 12 ########
#> ######Reconstruction 13 ########
#> ######Reconstruction 14 ########
#> ######Reconstruction 15 ########
#> ######Reconstruction 16 ########
#> ######Reconstruction 17 ########
#> ######Reconstruction 18 ########
#> ######Reconstruction 19 ########
#> ######Reconstruction 20 ########
#> ######Reconstruction 21 ########
#> ######Reconstruction 22 ########
#> ######Reconstruction 23 ########
#> ######Reconstruction 24 ########
#> ######Reconstruction 25 ########
#> ######Reconstruction 26 ########
#> ######Reconstruction 27 ########
#> ######Reconstruction 28 ########
#> ######Reconstruction 29 ########
#> ######Reconstruction 30 ########
#> ######Reconstruction 31 ########
#> ######Reconstruction 32 ########
#> ######Reconstruction 33 ########
#> ######Reconstruction 34 ########
#> ######Reconstruction 35 ########
#> ######Reconstruction 36 ########
#> ######Reconstruction 37 ########
#> ######Reconstruction 38 ########
#> ######Reconstruction 39 ########
#> ######Reconstruction 40 ########
#> ######Reconstruction 41 ########
#> ######Reconstruction 42 ########
#> ######Reconstruction 43 ########
#> ######Reconstruction 44 ########
#> ######Reconstruction 45 ########
#> ######Reconstruction 46 ########
#> ######Reconstruction 47 ########
#> ######Reconstruction 48 ########
#> ######Reconstruction 49 ########
#> ######Reconstruction 50 ########
#> ######Reconstruction 51 ########
#> ######Reconstruction 52 ########
#> ######Reconstruction 53 ########
#> ######Reconstruction 54 ########
#> ######Reconstruction 55 ########
#> ######Reconstruction 56 ########
#> ######Reconstruction 57 ########
#> ######Reconstruction 58 ########
#> ######Reconstruction 59 ########
#> ######Reconstruction 60 ########
#> ######Reconstruction 61 ########
#> ######Reconstruction 62 ########
#> ######Reconstruction 63 ########
#> ######Reconstruction 64 ########
#> ######Reconstruction 65 ########
#> ######Reconstruction 66 ########
#> ######Reconstruction 67 ########
#> ######Reconstruction 68 ########
#> ######Reconstruction 69 ########
#> ######Reconstruction 70 ########
#> ######Reconstruction 71 ########
#> ######Reconstruction 72 ########
#> ######Reconstruction 73 ########
#> ######Reconstruction 74 ########
#> ######Reconstruction 75 ########
#> ######Reconstruction 76 ########
#> ######Reconstruction 77 ########
#> ######Reconstruction 78 ########
#> ######Reconstruction 79 ########
#> ######Reconstruction 80 ########
#> ######Reconstruction 81 ########
#> ######Reconstruction 82 ########
#> ######Reconstruction 83 ########
#> ######Reconstruction 84 ########
#> ######Reconstruction 85 ########
#> ######Reconstruction 86 ########
#> ######Reconstruction 87 ########
#> ######Reconstruction 88 ########
#> ######Reconstruction 89 ########
#> ######Reconstruction 90 ########
#> ######Reconstruction 91 ########
#> ######Reconstruction 92 ########
#> ######Reconstruction 93 ########
#> ######Reconstruction 94 ########
#> ######Reconstruction 95 ########
#> ######Reconstruction 96 ########
#> ######Reconstruction 97 ########
#> ######Reconstruction 98 ########
#> ######Reconstruction 99 ########
#> ######Reconstruction 100 ########
#> ######Reconstruction 101 ########
#> ######Reconstruction 102 ########
#> ######Reconstruction 103 ########
#> ######Reconstruction 104 ########
#> ######Reconstruction 105 ########
#> ######Reconstruction 106 ########
#> ######Reconstruction 107 ########
#> ######Reconstruction 108 ########
#> ######Reconstruction 109 ########
#> ######Reconstruction 110 ########
#> ######Reconstruction 111 ########
#> ######Reconstruction 112 ########
#> ######Reconstruction 113 ########
#> ######Reconstruction 114 ########
#> ######Reconstruction 115 ########
#> ######Reconstruction 116 ########
#> ######Reconstruction 117 ########
#> ######Reconstruction 118 ########
#> ######Reconstruction 119 ########
#> ######Reconstruction 120 ########
#> ######Reconstruction 121 ########
#> ######Reconstruction 122 ########
#> ######Reconstruction 123 ########
#> ######Reconstruction 124 ########
#> ######Reconstruction 125 ########
#> ######Reconstruction 126 ########
#> ######Reconstruction 127 ########
#> ######Reconstruction 128 ########
#> ######Reconstruction 129 ########
#> ######Reconstruction 130 ########
#> ######Reconstruction 131 ########
#> ######Reconstruction 132 ########
#> ######Reconstruction 133 ########
#> ######Reconstruction 134 ########
#> ######Reconstruction 135 ########
#> ######Reconstruction 136 ########
#> ######Reconstruction 137 ########
#> ######Reconstruction 138 ########
#> ######Reconstruction 139 ########
#> ######Reconstruction 140 ########
#> ######Reconstruction 141 ########
#> ######Reconstruction 142 ########
#> ######Reconstruction 143 ########
#> ######Reconstruction 144 ########
#> ######Reconstruction 145 ########
#> ######Reconstruction 146 ########
#> ######Reconstruction 147 ########
#> ######Reconstruction 148 ########
#> ######Reconstruction 149 ########
#> ######Reconstruction 150 ########
#> ######Reconstruction 151 ########
#> ######Reconstruction 152 ########
#> ######Reconstruction 153 ########
#> ######Reconstruction 154 ########
#> ######Reconstruction 155 ########
#> ######Reconstruction 156 ########
#> ######Reconstruction 157 ########
#> ######Reconstruction 158 ########
#> ######Reconstruction 159 ########
#> ######Reconstruction 160 ########
#> ######Reconstruction 161 ########
#> ######Reconstruction 162 ########
#> ######Reconstruction 163 ########
#> ######Reconstruction 164 ########
#> ######Reconstruction 165 ########
#> ######Reconstruction 166 ########
#> ######Reconstruction 167 ########
#> ######Reconstruction 168 ########
#> ######Reconstruction 169 ########
#> ######Reconstruction 170 ########
#> ######Reconstruction 171 ########
#> ######Reconstruction 172 ########
#> ######Reconstruction 173 ########
#> ######Reconstruction 174 ########
#> ######Reconstruction 175 ########
#> ######Reconstruction 176 ########
#> ######Reconstruction 177 ########
#> ######Reconstruction 178 ########
#> ######Reconstruction 179 ########
#> ######Reconstruction 180 ########
#> ######Reconstruction 181 ########
#> ######Reconstruction 182 ########
#> ######Reconstruction 183 ########
#> ######Reconstruction 184 ########
#> ######Reconstruction 185 ########
#> ######Reconstruction 186 ########
#> ######Reconstruction 187 ########
#> ######Reconstruction 188 ########
#> ######Reconstruction 189 ########
#> ######Reconstruction 190 ########
#> ######Reconstruction 191 ########
#> ######Reconstruction 192 ########
#> ######Reconstruction 193 ########
#> ######Reconstruction 194 ########
#> ######Reconstruction 195 ########
#> ######Reconstruction 196 ########
#> ######Reconstruction 197 ########
#> ######Reconstruction 198 ########
#> ######Reconstruction 199 ########
#> ######Reconstruction 200 ########
#> ######Reconstruction 201 ########
#> ######Reconstruction 202 ########
#> ######Reconstruction 203 ########
#> ######Reconstruction 204 ########
#> ######Reconstruction 205 ########
#> ######Reconstruction 206 ########
#> ######Reconstruction 207 ########
#> ######Reconstruction 208 ########
#> ######Reconstruction 209 ########
#> ######Reconstruction 210 ########
#> ######Reconstruction 211 ########
#> ######Reconstruction 212 ########
#> ######Reconstruction 213 ########
#> ######Reconstruction 214 ########
#> ######Reconstruction 215 ########
#> ######Reconstruction 216 ########
#> ######Reconstruction 217 ########
#> ######Reconstruction 218 ########
#> ######Reconstruction 219 ########
#> ######Reconstruction 220 ########
#> ######Reconstruction 221 ########
#> ######Reconstruction 222 ########
#> ######Reconstruction 223 ########
#> ######Reconstruction 224 ########
#> ######Reconstruction 225 ########
#> ######Reconstruction 226 ########
#> ######Reconstruction 227 ########
#> ######Reconstruction 228 ########
#> ######Reconstruction 229 ########
#> ######Reconstruction 230 ########
#> ######Reconstruction 231 ########
#> ######Reconstruction 232 ########
#> ######Reconstruction 233 ########
#> ######Reconstruction 234 ########
#> ######Reconstruction 235 ########
#> ######Reconstruction 236 ########
#> ######Reconstruction 237 ########
#> ######Reconstruction 238 ########
#> ######Reconstruction 239 ########
#> ######Reconstruction 240 ########
#> ######Reconstruction 241 ########
#> ######Reconstruction 242 ########
#> ######Reconstruction 243 ########
#> ######Reconstruction 244 ########
#> ######Reconstruction 245 ########
#> ######Reconstruction 246 ########
#> ######Reconstruction 247 ########
#> ######Reconstruction 248 ########
#> ######Reconstruction 249 ########
#> ######Reconstruction 250 ########
#> ######Reconstruction 251 ########
#> ######Reconstruction 252 ########
#> ######Reconstruction 253 ########
#> ######Reconstruction 254 ########
#> ######Reconstruction 255 ########
#> ######Reconstruction 256 ########
#> ######Reconstruction 257 ########
#> ######Reconstruction 258 ########
#> ######Reconstruction 259 ########
#> ######Reconstruction 260 ########
#> ######Reconstruction 261 ########
#> ######Reconstruction 262 ########
#> ######Reconstruction 263 ########
#> ######Reconstruction 264 ########
#> ######Reconstruction 265 ########
#> ######Reconstruction 266 ########
#> ######Reconstruction 267 ########
#> ######Reconstruction 268 ########
#> ######Reconstruction 269 ########
#> ######Reconstruction 270 ########
#> ######Reconstruction 271 ########
#> ######Reconstruction 272 ########
#> ######Reconstruction 273 ########
#> ######Reconstruction 274 ########
#> ######Reconstruction 275 ########
#> ######Reconstruction 276 ########
#> ######Reconstruction 277 ########
#> ######Reconstruction 278 ########
#> ######Reconstruction 279 ########
#> ######Reconstruction 280 ########
#> ######Reconstruction 281 ########
#> ######Reconstruction 282 ########
#> ######Reconstruction 283 ########
#> ######Reconstruction 284 ########
#> ######Reconstruction 285 ########
#> ######Reconstruction 286 ########
#> ######Reconstruction 287 ########
#> ######Reconstruction 288 ########
#> ######Reconstruction 289 ########
#> ######Reconstruction 290 ########
#> ######Reconstruction 291 ########
#> ######Reconstruction 292 ########
#> ######Reconstruction 293 ########
#> ######Reconstruction 294 ########
#> ######Reconstruction 295 ########
#> ######Reconstruction 296 ########
#> ######Reconstruction 297 ########
#> ######Reconstruction 298 ########
#> ######Reconstruction 299 ########
#> ######Reconstruction 300 ########
#> ######Reconstruction 301 ########
#> ######Reconstruction 302 ########
#> ######Reconstruction 303 ########
#> ######Reconstruction 304 ########
#> ######Reconstruction 305 ########
#> ######Reconstruction 306 ########
#> ######Reconstruction 307 ########
#> ######Reconstruction 308 ########
#> ######Reconstruction 309 ########
#> ######Reconstruction 310 ########
#> ######Reconstruction 311 ########
#> ######Reconstruction 312 ########
#> ######Reconstruction 313 ########
#> ######Reconstruction 314 ########
#> ######Reconstruction 315 ########
#> ######Reconstruction 316 ########
#> ######Reconstruction 317 ########
#> ######Reconstruction 318 ########
#> ######Reconstruction 319 ########
#> ######Reconstruction 320 ########
#> ######Reconstruction 321 ########
#> ######Reconstruction 322 ########
#> ######Reconstruction 323 ########
#> ######Reconstruction 324 ########
#> ######Reconstruction 325 ########
#> ######Reconstruction 326 ########
#> ######Reconstruction 327 ########
#> ######Reconstruction 328 ########
#> ######Reconstruction 329 ########
#> ######Reconstruction 330 ########
#> ######Reconstruction 331 ########
#> ######Reconstruction 332 ########
#> ######Reconstruction 333 ########
#> ######Reconstruction 334 ########
#> ######Reconstruction 335 ########
#> ######Reconstruction 336 ########
#> ######Reconstruction 337 ########
#> ######Reconstruction 338 ########
#> ######Reconstruction 339 ########
#> ######Reconstruction 340 ########
#> ######Reconstruction 341 ########
#> ######Reconstruction 342 ########
#> ######Reconstruction 343 ########
#> ######Reconstruction 344 ########
#> ######Reconstruction 345 ########
#> ######Reconstruction 346 ########
#> ######Reconstruction 347 ########
#> ######Reconstruction 348 ########
#> ######Reconstruction 349 ########
#> ######Reconstruction 350 ########
#> ######Reconstruction 351 ########
#> ######Reconstruction 352 ########
#> ######Reconstruction 353 ########
#> ######Reconstruction 354 ########
#> ######Reconstruction 355 ########
#> ######Reconstruction 356 ########
#> ######Reconstruction 357 ########
#> ######Reconstruction 358 ########
#> ######Reconstruction 359 ########
#> ######Reconstruction 360 ########
#> ######Reconstruction 361 ########
#> ######Reconstruction 362 ########
#> ######Reconstruction 363 ########
#> ######Reconstruction 364 ########
#> ######Reconstruction 365 ########
#> ######Reconstruction 366 ########
#> ######Reconstruction 367 ########
#> ######Reconstruction 368 ########
#> ######Reconstruction 369 ########
#> ######Reconstruction 370 ########
#> ######Reconstruction 371 ########
#> ######Reconstruction 372 ########
#> ######Reconstruction 373 ########
#> ######Reconstruction 374 ########
#> ######Reconstruction 375 ########
#> ######Reconstruction 376 ########
#> ######Reconstruction 377 ########
#> ######Reconstruction 378 ########
#> ######Reconstruction 379 ########
#> ######Reconstruction 380 ########
#> ######Reconstruction 381 ########
#> ######Reconstruction 382 ########
#> ######Reconstruction 383 ########
#> ######Reconstruction 384 ########
#> ######Reconstruction 385 ########
#> ######Reconstruction 386 ########
#> ######Reconstruction 387 ########
#> ######Reconstruction 388 ########
#> ######Reconstruction 389 ########
#> ######Reconstruction 390 ########
#> ######Reconstruction 391 ########
#> ######Reconstruction 392 ########
#> ######Reconstruction 393 ########
#> ######Reconstruction 394 ########
#> ######Reconstruction 395 ########
#> ######Reconstruction 396 ########
#> ######Reconstruction 397 ########
#> ######Reconstruction 398 ########
#> ######Reconstruction 399 ########
#> ######Reconstruction 400 ########
#> ######Reconstruction 401 ########
#> ######Reconstruction 402 ########
#> ######Reconstruction 403 ########
#> ######Reconstruction 404 ########
#> ######Reconstruction 405 ########
#> ######Reconstruction 406 ########
#> ######Reconstruction 407 ########
#> ######Reconstruction 408 ########
#> ######Reconstruction 409 ########
#> ######Reconstruction 410 ########
#> ######Reconstruction 411 ########
#> ######Reconstruction 412 ########
#> ######Reconstruction 413 ########
#> ######Reconstruction 414 ########
#> ######Reconstruction 415 ########
#> ######Reconstruction 416 ########
#> ######Reconstruction 417 ########
#> ######Reconstruction 418 ########
#> ######Reconstruction 419 ########
#> ######Reconstruction 420 ########
#> ######Reconstruction 421 ########
#> ######Reconstruction 422 ########
#> ######Reconstruction 423 ########
#> ######Reconstruction 424 ########
#> ######Reconstruction 425 ########
#> ######Reconstruction 426 ########
#> ######Reconstruction 427 ########
#> ######Reconstruction 428 ########
#> ######Reconstruction 429 ########
#> ######Reconstruction 430 ########
#> ######Reconstruction 431 ########
#> ######Reconstruction 432 ########
#> ######Reconstruction 433 ########
#> ######Reconstruction 434 ########
#> ######Reconstruction 435 ########
#> ######Reconstruction 436 ########
#> ######Reconstruction 437 ########
#> ######Reconstruction 438 ########
#> ######Reconstruction 439 ########
#> ######Reconstruction 440 ########
#> ######Reconstruction 441 ########
#> ######Reconstruction 442 ########
#> ######Reconstruction 443 ########
#> ######Reconstruction 444 ########
#> ######Reconstruction 445 ########
#> ######Reconstruction 446 ########
#> ######Reconstruction 447 ########
#> ######Reconstruction 448 ########
#> ######Reconstruction 449 ########
#> ######Reconstruction 450 ########
#> ######Reconstruction 451 ########
#> ######Reconstruction 452 ########
#> ######Reconstruction 453 ########
#> ######Reconstruction 454 ########
#> ######Reconstruction 455 ########
#> ######Reconstruction 456 ########
#> ######Reconstruction 457 ########
#> ######Reconstruction 458 ########
#> ######Reconstruction 459 ########
#> ######Reconstruction 460 ########
#> ######Reconstruction 461 ########
#> ######Reconstruction 462 ########
#> ######Reconstruction 463 ########
#> ######Reconstruction 464 ########
#> ######Reconstruction 465 ########
#> ######Reconstruction 466 ########
#> ######Reconstruction 467 ########
#> ######Reconstruction 468 ########
#> ######Reconstruction 469 ########
#> ######Reconstruction 470 ########
#> ######Reconstruction 471 ########
#> ######Reconstruction 472 ########
#> ######Reconstruction 473 ########
#> ######Reconstruction 474 ########
#> ######Reconstruction 475 ########
#> ######Reconstruction 476 ########
#> ######Reconstruction 477 ########
#> ######Reconstruction 478 ########
#> ######Reconstruction 479 ########
#> ######Reconstruction 480 ########
#> ######Reconstruction 481 ########
#> ######Reconstruction 482 ########
#> ######Reconstruction 483 ########
#> ######Reconstruction 484 ########
#> ######Reconstruction 485 ########
#> ######Reconstruction 486 ########
#> ######Reconstruction 487 ########
#> ######Reconstruction 488 ########
#> ######Reconstruction 489 ########
#> ######Reconstruction 490 ########
#> ######Reconstruction 491 ########
#> ######Reconstruction 492 ########
#> ######Reconstruction 493 ########
#> ######Reconstruction 494 ########
#> ######Reconstruction 495 ########
#> ######Reconstruction 496 ########
#> ######Reconstruction 497 ########
#> ######Reconstruction 498 ########
#> ######Reconstruction 499 ########
#> ######Reconstruction 500 ########