TIGER filtering

This function returns a object of class 'tigerFilt' that has been filtered to ensure TIGER batch correction can run successfully

tiger_filter(omicsData, sampletype_cname, test_val)

Arguments

omicsData

an object of the class 'pepData', 'proData', 'metabData', 'lipidData', or 'nmrData', usually created by as.pepData, as.proData, as.metabData, as.lipidData, or as.nmrData, respectively.

sampletype_cname

character string giving the name of the column in omicsData$f_data that contains the sample type information (such as quality control samples)

test_val

character string giving the name of the value within the column sampletype_cname to be used as the testing value for TIGER

Value

Object of tigerFilt class

Author

Damon Leach

Examples

library(malbacR)
library(pmartRdata)
data(pmart_amide)
pmart_amide <- group_designation(pmart_amide,main_effects = "group",batch_id = "batch")
tigerFilt <- tiger_filter(omicsData = pmart_amide,sampletype_cname = "group", test_val = "QC")
pmart_amide_filt <- apply_tigerFilt(filter_object = tigerFilt,omicsData = pmart_amide)
#> There are no molecules that need to be removed. Returning the original omicsData
amide_tiger <- bc_tiger(omicsData = pmart_amide_filt, sampletype_cname = "group", test_val = "QC")
#> + Initialising...   2023-10-17 12:28:06
#> + Computing target values...   2023-10-17 12:28:06
#> + Selecting highly-correlated variables...   2023-10-17 12:28:06
#>   - Computing correlation coefficients...
#>   - Selecting variables...
#> + Data correction started.   2023-10-17 12:28:07
#>   |                                                                      | 0 % ~calculating    |==                                                                    | 2 % ~02m 55s        |===                                                                   | 4 % ~02m 55s        |====                                                                  | 5 % ~02m 55s        |=====                                                                 | 7 % ~02m 53s        |=======                                                               | 9 % ~02m 50s        |========                                                              | 11% ~02m 46s        |=========                                                             | 12% ~02m 42s        |==========                                                            | 14% ~02m 38s        |============                                                          | 16% ~02m 34s        |=============                                                         | 18% ~02m 31s        |==============                                                        | 20% ~02m 27s        |===============                                                       | 21% ~02m 25s        |=================                                                     | 23% ~02m 22s        |==================                                                    | 25% ~02m 18s        |===================                                                   | 27% ~02m 14s        |====================                                                  | 29% ~02m 10s        |======================                                                | 30% ~02m 06s        |=======================                                               | 32% ~02m 03s        |========================                                              | 34% ~01m 60s        |=========================                                             | 36% ~01m 56s        |===========================                                           | 38% ~01m 53s        |============================                                          | 39% ~01m 49s        |=============================                                         | 41% ~01m 46s        |==============================                                        | 43% ~01m 42s        |================================                                      | 45% ~01m 39s        |=================================                                     | 46% ~01m 35s        |==================================                                    | 48% ~01m 32s        |===================================                                   | 50% ~01m 29s        |=====================================                                 | 52% ~01m 25s        |======================================                                | 54% ~01m 22s        |=======================================                               | 55% ~01m 19s        |========================================                              | 57% ~01m 16s        |==========================================                            | 59% ~01m 12s        |===========================================                           | 61% ~01m 09s        |============================================                          | 62% ~01m 06s        |=============================================                         | 64% ~01m 03s        |===============================================                       | 66% ~60s            |================================================                      | 68% ~56s            |=================================================                     | 70% ~53s            |==================================================                    | 71% ~50s            |====================================================                  | 73% ~47s            |=====================================================                 | 75% ~44s            |======================================================                | 77% ~41s            |=======================================================               | 79% ~38s            |=========================================================             | 80% ~35s            |==========================================================            | 82% ~31s            |===========================================================           | 84% ~28s            |============================================================          | 86% ~25s            |==============================================================        | 88% ~22s            |===============================================================       | 89% ~19s            |================================================================      | 91% ~16s            |=================================================================     | 93% ~13s            |===================================================================   | 95% ~10s            |====================================================================  | 96% ~06s            |===================================================================== | 98% ~03s            |======================================================================| 100% elapsed=02m 59s
#>   - Merging results...
#> + Completed.   2023-10-17 12:31:07