Perform scale method on data

This function is a modified version of the function ScalingMethods from the package DiffCorr, but it allows for missing values for Power and Range scaling

Usage

scale_method(
  data,
  methods = c("auto", "range", "pareto", "vast", "level", "power")
)

Arguments

data

a data matrix or data frame object where each row is a biomolecule and each column each sample or replicate

methods

a string character specifying the pre-treatment method and can be "auto", "range", "pareto", "vast", "level", or "power"

Value

a data matrix or data frame object (depending on what was the input) that has undergone pre-treatment method

Author

Damon Leach

Examples

library(malbacR)
data(pmart_amide)
malbacR:::scale_method(pmart_amide$e_data[,-1])
#>          QC1        A2         A3        A4        A5        A6           A7
#> 15 0.3229682 0.3848103 -0.1641426 -1.208995 -1.250579 0.7929884 -0.006639176
#>            A8        A9       QC2       A11         A12       A13        A14
#> 15 0.03505824 -1.502804 -1.081568 -1.142301 -0.05897936 -1.405276 -0.2025297
#>          A15       A16       A17         A18         QC3        A20       A21
#> 15 0.6792133 -1.380232 -0.298431 -0.01278994 -0.05046975 0.07133706 0.5234411
#>           A22       A23       A24       A25        A26       A27       QC4
#> 15 -0.2342473 -1.248139 0.4196079 -1.276341 -0.3761048 -1.449013 0.8270291
#>         A29       A30      A31       A32       A33      A34       A35       A36
#> 15 -1.34525 -1.339735 1.132264 -1.303362 0.9252732 1.049954 0.3844769 -1.270507
#>          QC5        A38        A39       A40       A41       A42      A43
#> 15 -1.342251 -0.1580015 -0.2839848 0.5691168 0.6345035 -1.248531 -1.36036
#>          A44       A45      QC6       A47        A48       A49         A50
#> 15 0.7616503 -1.558294 -1.38518 -1.106793 0.07066937 -1.391756 -0.08324261
#>         A51        A52       A53       A54        QC7       A56       A57
#> 15 -1.13747 -0.1608431 -1.356001 0.2396596 -0.1266313 -1.238318 -0.184127
#>          A58        A59       A60         A61       A62       A63       QC8
#> 15 -1.320453 -0.4278337 -1.443187 -0.07960773 -1.182377 -1.450614 0.8598588
#>         A65       A66       A67        A68      A69       A70       A71
#> 15 0.296197 -1.353368 0.5427972 -0.3605241 -1.52238 -1.460302 0.5290859
#>          A72        QC9        A74        A75       A76       A77        A78
#> 15 -1.414705 -0.2691974 0.03779785 -0.5389647 0.4205004 -1.447128 -0.4823557
#>          A79       A80       A81       QC10       A83       A84       A85
#> 15 0.5926973 -0.143983 0.4105563 -0.4039379 0.5222731 -1.483661 -1.397724
#>           A86      A87       A88       A89       A90       QC11        A92
#> 15 -0.4761701 0.230426 -1.505632 -1.388901 -1.369288 -0.5832279 -0.4466349
#>          A93       A94       A95       A96        A97        A98       A99
#> 15 -1.424673 -1.554382 0.4953775 -1.494657 -0.3497733 -0.6295169 -1.397411
#>          QC12      A101      A102       A103       A104       A105      A106
#> 15 -0.6193879 -1.588029 -1.546415 -0.7567127 -0.6589706 -0.6693968 -1.485752
#>          A107       A108       QC13       A110       A111       A112       A113
#> 15 -0.7116108 -0.5810146 -0.3401062 -0.6040067 -0.5952132 -0.6847994 -0.9419115
#>          A114       A115     A116       A117      QC14      A119       A120
#> 15 -0.7248884 -0.1299021 -0.49381 0.07626492 -1.579068 -1.603621 -0.9907493
#>          A121       A122       A123      A124      A125       A126       QC15
#> 15 -0.8186879 -0.6183104 -0.9725748 -0.567622 -1.426449 -0.6247686 -0.5794946
#>          A128       A129       A130     A131      A132       A133       A134
#> 15 -0.6282773 -0.6035843 -0.7745532 -1.32831 -0.221385 -0.5419883 -0.9358234
#>         A135      QC16      A137       A138       A139      A140        A141
#> 15 -0.371227 0.2524853 -1.564519 -0.4003489 -0.1812378 -0.356987 -0.01038253
#>          A142       A143      A144       QC17       A146       A147      A148
#> 15 -0.4338574 -0.2637492 -1.513712 -0.3663788 -0.8047072 -0.6724489 -0.659728
#>          A149     A150       A151       A152      A153      QC18       A155
#> 15 -0.9236893 -0.80914 -0.4039122 -0.7875478 -1.748879 -1.576349 -0.3668892
#>          A156      A157      A158    A159      A160       A161       A162
#> 15 -0.2555499 -1.613464 -0.369605 -0.7263 -0.473309 -0.3612382 -0.5634606
#>          QC19       A164       A165       A166       A167       A168       A169
#> 15 -0.3928002 -0.4145754 -0.4612213 -0.5109207 -0.7176201 -0.9790066 -0.7608275
#>          A170       A171   QC20      A173       A174       A175       A176
#> 15 -0.7616251 -0.7938758 -0.428 -0.716008 -0.1755485 -0.3800628 -0.7429369
#>         A177       A178       A179       A180       QC21       A182       A183
#> 15 -0.357579 -0.8260298 -0.4114034 -0.4284366 -0.5386994 -0.7747461 -0.7174992
#>          A184       A185      A186       A187      A188      A189       QC22
#> 15 -0.9646946 -0.9585036 -0.802017 -0.8390005 -0.499387 -1.804001 -0.4564943
#>          A191       A192      A193       A194       A195       A196      A197
#> 15 -0.7506891 -0.7816862 -1.011543 -0.5284711 -0.9574033 -0.3681749 -1.069488
#>          A198       QC23       A200      A201       A202       A203       A204
#> 15 -0.9345311 -0.3337995 -0.8027805 -1.552873 -0.9029458 -0.4473823 -0.7931951
#>         A205       A206      A207       QC24     A209       A210      A211
#> 15 -0.792117 -0.9856603 -1.138059 -0.5964213 -0.95679 -0.9737947 -1.147316
#>          A212       A213        A214       A215       A216       QC25     QC29
#> 15 -0.8190392 -0.2034341 -0.06546176 -0.7428168 -0.8719465 -0.3406625 3.608932
#>        A246    A247     A248     A249     A250      A251     A252     A253
#> 15 1.056015 1.13151 1.767481 1.621106 1.106414 0.3922361 3.751123 4.121252
#>        QC30      A255     A256     A257     A258     A259     A260     A261
#> 15 2.367244 0.9626347 1.930545 1.842779 1.807894 2.263433 1.664901 2.199261
#>        A262      QC31      A264     A265     A266     A267     A268     A269
#> 15 1.506892 0.4698756 0.7795432 1.430996 2.103319 1.485374 1.754094 2.077355
#>        A270    A271     QC32     A273      A274     A275      A276     A277
#> 15 2.171684 2.07277 -2.24164 1.276991 0.7917703 1.128321 0.4761253 3.026718
#>         A278      A279     A280       QC33       A282     A283      A284
#> 15 0.9216729 0.6951876 1.006225 -0.3533505 -0.3365325 2.695388 0.7759253
#>        A285     A286      A287       A288      A289     QC34      A291
#> 15 1.200949 3.182314 0.7945863 -0.4752438 0.4686687 1.141985 0.8816403
#>         A292     A293       A294       A295        A296      A297      A298
#> 15 0.6219556 1.709916 -0.0801751 -0.2879536 -0.05036136 0.7025476 0.8289996
#>          QC35     A300     A301      A302      A303      A304   A305      A306
#> 15 -0.3772392 1.272737 2.556838 0.9251838 0.8849261 0.2441716 0.4307 0.7744237
#>         A307      QC36      A309       A310      A311      A312     A313
#> 15 0.9407798 0.2658754 0.7475056 0.04858607 0.7971057 0.5951324 1.687408
#>        A314      A315       A316     QC37     A318     A319    A320     A321
#> 15 1.680233 0.1319362 -0.9288164 1.106623 1.652216 1.261384 1.18153 0.754079
#>         A322       A323      A324      A325      QC38     A327      A328
#> 15 0.8658622 -0.1155786 0.8682973 0.8906228 0.3144961 1.030002 0.6000971
#>         A329     A330      A331      A332      A333      A334     QC39     A336
#> 15 0.6758801 3.017398 0.7374677 0.6519909 0.5676678 0.6805467 1.060029 1.316549
#>         A337      A338        A339       A340     A341       A342       A343
#> 15 0.9717903 0.5955378 -0.08240923 -0.3943981 0.367872 -0.5252502 -0.1226674
#>        QC40     A345    A346        A347      A348     A349      A350     A351
#> 15 1.305455 0.526435 1.11402 -0.05825587 0.4918337 1.057673 0.7344481 0.983067
#>         A352      QC41      A354      A355     A356      A357     A358
#> 15 0.6148134 0.1047315 0.3460604 0.6763899 2.562207 0.4004878 2.203078
#>         A359      A360     A361     QC42     A363     A364      A365      A366
#> 15 0.6068776 0.5761286 2.335616 2.029879 1.199875 1.283106 0.1064318 0.4693652
#>         A367       A368      A369      A370       QC43     A372       A373
#> 15 0.2967467 0.07057505 0.4750962 0.6174247 -0.7745964 1.364124 -0.3774718
#>        A374      A375      A376     A377       A378      A379       QC44
#> 15 2.406738 0.4416508 0.3660014 0.326997 0.03722721 0.4528392 0.06886356
#>        A381     A382      A383     A384      A385      A386     A387      A388
#> 15 1.269875 1.249927 0.4049785 0.415407 0.5741722 0.2205221 2.424574 0.4403686
#>         QC45       A390     A391     A392       A393      A394      A395
#> 15 0.8838727 -0.5329867 1.424337 1.608571 -0.7892137 -0.499366 0.7035668
#>         A396     A397       QC46     A399      A400     A401     A402     A403
#> 15 -0.395381 1.354362 -0.4796359 1.655801 0.8657854 1.635314 1.507103 1.836816
#>        A404      A405       A406      QC47      A408       A409         A410
#> 15 1.419743 0.1000136 -0.5509436 0.2055878 0.9762379 -0.6505916 -0.004394719
#>           A411     A412        A413      A414       A415      QC48      A417
#> 15 -0.05225492 1.327686 -0.02180781 0.8869234 -0.5801356 0.9333976 0.9110512
#>          A418        A419      A420        A421        A422      A423      A424
#> 15 -0.2952847 -0.05968004 0.9141186 -0.07958328 -0.05457079 0.6310012 0.9726111
#>         QC49      A426     A427      A428       A429     A430      A431
#> 15 0.9197356 0.5905353 1.931396 0.2225812 -0.3524973 0.723341 0.8017667
#>           A432       A433      QC50       A219      A220       A221     A222
#> 15 -0.03077782 -0.1631553 0.6164909 -0.9156253 -1.706454 -0.2742242 0.337176
#>          A223       A224        A225      A226      QC51     A228      A229
#> 15 -0.2261085 -0.4302167 -0.04732196 -1.095381 0.2668558 0.630247 0.5448096
#>          A230       A231       A232      A233      A234      A235       QC52
#> 15 -0.2896985 -0.1840293 -0.1956523 0.1811862 0.7554334 -0.805407 -0.2163363
#>         A237      A238      A239     A240      A241      A242    A243     A244
#> 15 -1.003466 0.9575587 0.3962451 1.889783 0.5290084 0.1392555 1.05328 2.049707
#>       QC53       QC54      A436       A437       A438       A439        A440
#> 15 1.13384 -0.4739801 0.6719797 -0.3244037 -0.2187528 -0.6069874 -0.02241263
#>          A441       A442       A443       QC55       A445       A446       A447
#> 15 -0.2022468 -0.4359903 -0.3766723 -0.3150089 -0.5675085 -0.4011313 -0.2517078
#>        A448       A449       A450       A451       A452       QC56       A454
#> 15 -0.28414 -0.6019218 -0.5624526 -0.4301137 -0.6966499 -0.5539601 -0.6079871
#>          A456       A457      A458       A459       A460       A461       QC57
#> 15 -0.4641358 -0.1431229 -0.406696 -0.7342996 -0.3922611 -0.6629408 -0.6240247
#>          A463       A464       A465       A466      A467      A468       A470
#> 15 -0.6909373 -0.7434035 -0.2943089 -0.6655028 0.1852875 -0.288329 -0.6558446
#>          QC58      A473        A474      A475        A476       A477      A478
#> 15 -0.3944596 0.4409525 -0.03278596 0.4777275 -0.09227154 -0.1382523 0.1480607
#>          A479       A481       A483       A484      A485       A486      A487
#> 15 -0.3226689 -0.7171554 -0.4070988 -0.6799089 -0.385794 -0.1653475 0.2244041
#>          A488        QC60     A490        A491       A492        A493
#> 15 -0.2472723 -0.04798103 0.188789 -0.03013801 -0.1359479 0.007535257
#>          A494       A495      A496         A497        QC61       A499
#> 15 0.02533058 -0.3783353 0.3594934 -0.004668153 -0.05750885 0.01901368
#>          A500       A501       A502       A503       A504      A505       A506
#> 15 -0.1253213 -0.5474963 -0.4676619 -0.2902086 -0.9885211 -1.022432 -0.8183834
#>         QC62       A508       A509       A510       A511       A512       A513
#> 15 -0.498218 -0.8074561 -0.5139302 -0.3615842 -0.3652489 -0.6032389 -0.1906447
#>          A514       A515       QC63    A517        A518       A519       A520
#> 15 -0.5004729 -0.4728993 -0.2766642 -0.4967 -0.03671985 -0.5346383 -0.5585818
#>          A521       A522       A523       A524       QC64       A526       A527
#> 15 -0.5871461 -0.7259208 -0.5002269 -0.8776396 -0.7409305 -0.1891032 -0.1843856
#>          A528       A529      A530       A531       A532       A533       QC65
#> 15 -0.3591394 -0.1084818 -0.562295 -0.3916757 -0.2162082 -0.9747433 0.06022756
#>          A535     A536       A537       A538       A540      A541       A542
#> 15 -0.1936175 5.000785 -0.2352449 -0.7086607 -0.8623499 -1.020697 -0.9271764
#>         QC66      A544      A545     A546       A547        A549    A550
#> 15 0.2867848 0.6657006 0.2246663 0.302616 -0.6099191 -0.08521451 1.59324
#>          A551       QC67       A554      A555      A556      A557       A558
#> 15 0.05480162 -0.0778833 0.01986572 0.6101913 0.4007971 0.1043134 -0.1009953
#>         A559      A560        QC68      A562      A563      A564        A565
#> 15 0.3400612 0.5762039 -0.01295513 0.1664854 0.5715028 0.2627128 -0.01461234
#>          A566      A567      A568       A569      QC69      A571        A572
#> 15 -0.2734016 0.3696088 0.4803943 0.03487505 0.1516163 0.2419782 -0.02904112
#>        A573        A574      A575      A576     A577       A578      QC70
#> 15 6.879784 -0.05206649 0.1542102 0.2993545 1.427673 -0.1810615 0.1044366
#>         A580      A581      A582      A583       A585       A586       A587
#> 15 0.2777454 -0.385325 -1.048308 -1.393605 -0.8209318 -0.8469292 -0.5755727
#>          A589      A590       A591       A592       A593       A594      A595
#> 15 -0.3018927 0.2115345 -0.5766752 -0.6825617 -0.1721022 -0.2606259 0.7369227
#>          A596       QC72        A598      A599      A600      A601        A602
#> 15 -0.4211858 -0.4508947 -0.05420123 0.3001256 0.4616904 0.2877493 -0.04820706
#>         A603       A604       A605        QC73      A607      A608       A609
#> 15 0.3359336 -0.1462937 -0.1778119 -0.06569509 0.0102201 -0.358598 -0.1522967
#>         A611       A612       A613      QC74       A616      A617       A618
#> 15 -0.330351 -0.8693554 -0.2579572 -2.323081 -0.2779753 0.0593015 0.01030997
#>          A619       A621       A622      A623      QC75      A625      A626
#> 15 -0.7811291 0.05193819 0.03225002 0.6887306 -0.240981 0.4048841 0.3577943
#>          A627      A628    A629       A630      A631       A632       QC76
#> 15 -0.9165381 0.3026907 4.16904 -0.3449489 -0.799691 -0.9733514 -0.4790549
#>          A634      A635       A636       A637      A638       A639       A640
#> 15 -0.5901435 0.5708328 -0.2554295 -0.2387815 0.2750638 -0.2456859 -0.1287687
#>         A641       QC77       A584        A610       A614     A620     A455
#> 15 0.2029579 0.03131434 -0.2877568 -0.05349843 -0.3647172 1.529059 1.192699
#>          A469       QC78       A472      A482       A539      A548       A553
#> 15 0.07469291 -0.8752306 -0.1282693 0.1127415 -0.5045548 -2.295151 -0.4501169
#>         QC79
#> 15 0.3599304
#>  [ reached 'max' / getOption("max.print") -- omitted 499 rows ]